+++ VoxSmith/src/VoxSmith/RayTracer/RayTracer.hpp
@@ -17,7 +17,7 @@ namespace VoxSmith
 	class VOX_SMITH_API RayTracer final
 	{
 	public:
-		RayTracer(const uint32_t width, const uint32_t height);
+		RayTracer(const float width, const float height);
 		~RayTracer() = default;
 
 		void trace(const glm::ivec3& dims, const ComputeShader& cShader) const;
